+++ services/gateway/src/main.ts
@@ -4,14 +4,14 @@ import errLogger from './tools/logger/logger';
 import Log from './tools/logger/log';
 import Broker from './broker';
 
-class App {
+export default class App {
   init(): void {
     const router = new Router();
     const broker = new Broker();
 
     try {
-      State.Router = router;
-      State.Broker = broker;
+      State.router = router;
+      State.broker = broker;
       router.init();
       broker.init();
     } catch (err) {
@@ -23,6 +23,12 @@ class App {
       router.close();
       broker.close();
     }
+  }
+
+  kill(): void {
+    State.router.close();
+    State.broker.close();
+    Log.log('Server', 'Server closed');
   }
 }