+++ services/gateway/src/tools/state.ts
@@ -1,8 +1,29 @@
+import type Broker from '../broker';
+import type Router from '../router';
 import type { IState } from '../types';
 
-const State: IState = {
-  Router: null,
-  Broker: null,
-};
+class State implements IState {
+  private _broker: Broker;
 
-export default State;
+  get broker(): Broker {
+    return this._broker;
+  }
+
+  set broker(value: Broker) {
+    this._broker = value;
+  }
+
+  private _router: Router;
+
+  get router(): Router {
+    return this._router;
+  }
+
+  set router(value: Router) {
+    this._router = value;
+  }
+}
+
+const state = new State();
+
+export default state;