+++ services/users/src/types/state.d.ts
@@ -8,4 +8,5 @@ export interface IConfigInterface {
   amqpURI: string;
   token: string;
   refToken: string;
+  mongoURI: string;
 }